Rust: Updated template project
Added two new tests that when set will check that the correct answer to
the full data is returned. These tests will help when trying to improve
a completed solution.
Moved all tests to a separate tests.rs file.
Any test with a expected value of 0 will not pass automatically. This is
so it is easier to see if a solution produced the correct value, since
the tests output wont be full of other test failures.